Move systems/trajectories to common/trajectories and lowercase its files
