+++ drake/common/trajectories/piecewise_function.h
