+++ drake/common/trajectories/test/piecewise_polynomial_test.cc
@@ -1,4 +1,4 @@
-#include "drake/systems/trajectories/PiecewisePolynomial.h"
+#include "drake/common/trajectories/piecewise_polynomial.h"
 
 #include <iostream>
 #include <random>
