+++ drake/common/trajectories/test/piecewise_polynomial_trajectory_test.cc
@@ -1,8 +1,8 @@
-#include "drake/systems/trajectories/piecewise_polynomial_trajectory.h"
+#include "drake/common/trajectories/piecewise_polynomial_trajectory.h"
 
 #include "gtest/gtest.h"
 
-#include "drake/systems/trajectories/PiecewisePolynomial.h"
+#include "drake/common/trajectories/piecewise_polynomial.h"
 
 using Eigen::MatrixXd;
