+++ drake/common/trajectories/trajectory.h
