+++ drake/solvers/trajectoryOptimization/direct_trajectory_optimization.h
@@ -5,9 +5,9 @@
 #include <Eigen/Core>
 
 #include "drake/common/drake_assert.h"
+#include "drake/common/trajectories/piecewise_polynomial.h"
+#include "drake/common/trajectories/piecewise_polynomial_trajectory.h"
 #include "drake/solvers/mathematical_program.h"
-#include "drake/systems/trajectories/PiecewisePolynomial.h"
-#include "drake/systems/trajectories/piecewise_polynomial_trajectory.h"
 
 namespace drake {
 namespace solvers {
