+++ drake/systems/CMakeLists.txt
@@ -2,6 +2,5 @@ add_subdirectory(analysis)
 add_subdirectory(framework)
 add_subdirectory(lcm)
 add_subdirectory(plants)
-add_subdirectory(trajectories)
 add_subdirectory(controllers)
 add_subdirectory(robotInterfaces)
