+++ drake/systems/framework/primitives/trajectory_source.h
@@ -2,10 +2,10 @@
 
 #include <cstdint>
 
+#include "drake/common/trajectories/piecewise_polynomial_trajectory.h"
 #include "drake/systems/framework/context.h"
 #include "drake/systems/framework/leaf_system.h"
 #include "drake/systems/framework/system_output.h"
-#include "drake/systems/trajectories/piecewise_polynomial_trajectory.h"
 
 namespace drake {
 namespace systems {
