+++ drake/systems/robotInterfaces/QPLocomotionPlan.h
@@ -9,13 +9,13 @@
 
 #include <lcm/lcm-cpp.hpp>
 
+#include "drake/common/trajectories/exponential_plus_piecewise_polynomial.h"
+#include "drake/common/trajectories/piecewise_polynomial.h"
 #include "drake/lcmt_qp_controller_input.hpp"
 #include "drake/multibody/rigid_body_tree.h"
 #include "drake/systems/controllers/zmpUtil.h"
 #include "drake/systems/robotInterfaces/BodyMotionData.h"
 #include "drake/systems/robotInterfaces/Side.h"
-#include "drake/systems/trajectories/PiecewisePolynomial.h"
-#include "drake/systems/trajectories/ExponentialPlusPiecewisePolynomial.h"
 
 class QuadraticLyapunovFunction {
   // TODO(tkoolen): move into its own file
